[util] Move ReflectionUtils to bookkeeper-common module
*Motivation*

RefelectionUtils contains helpers for loading classes from relections.
It is useful across the project. However the class is in `bookkeeper-server` module only.

*Changes*

Move this class to `bookkeeper-common` module
