+++ bookkeeper-common/src/main/java/org/apache/bookkeeper/common/util/ReflectionUtils.java
@@ -1,5 +1,4 @@
 /*
- *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -16,9 +15,8 @@
  * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
- *
  */
-package org.apache.bookkeeper.util;
+package org.apache.bookkeeper.common.util;
 
 import java.lang.reflect.Constructor;
 import java.util.Map;
