+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/auth/AuthProviderFactoryFactory.java
@@ -22,12 +22,11 @@ package org.apache.bookkeeper.auth;
 
 import java.io.IOException;
 import org.apache.bookkeeper.client.BKException;
+import org.apache.bookkeeper.common.util.ReflectionUtils;
 import org.apache.bookkeeper.conf.ClientConfiguration;
 import org.apache.bookkeeper.conf.ServerConfiguration;
 import org.apache.bookkeeper.proto.BookieConnectionPeer;
 import org.apache.bookkeeper.proto.ClientConnectionPeer;
-import org.apache.bookkeeper.util.ReflectionUtils;
-
 
 /**
  * A factory to manage the authentication provider factories.
