+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/LedgerStorageFactory.java
@@ -22,7 +22,7 @@ package org.apache.bookkeeper.bookie;
 
 import java.io.IOException;
 
-import org.apache.bookkeeper.util.ReflectionUtils;
+import org.apache.bookkeeper.common.util.ReflectionUtils;
 
 /**
  * A factory that creates {@link LedgerStorage} by reflection.
