+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/server/service/StatsProviderService.java
@@ -19,11 +19,11 @@
 package org.apache.bookkeeper.server.service;
 
 import java.io.IOException;
+import org.apache.bookkeeper.common.util.ReflectionUtils;
 import org.apache.bookkeeper.server.component.ServerLifecycleComponent;
 import org.apache.bookkeeper.server.conf.BookieConfiguration;
 import org.apache.bookkeeper.stats.NullStatsLogger;
 import org.apache.bookkeeper.stats.StatsProvider;
-import org.apache.bookkeeper.util.ReflectionUtils;
 
 /**
  * A {@link org.apache.bookkeeper.common.component.LifecycleComponent} that runs stats provider.
