+++ tests/shaded/distributedlog-core-shaded-test/src/test/java/org/apache/bookkeeper/tests/shaded/DistributedLogCoreShadedJarTest.java
@@ -30,6 +30,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 
+import dlshade.org.apache.bookkeeper.common.util.ReflectionUtils;
 import dlshade.org.apache.bookkeeper.conf.AbstractConfiguration;
 import dlshade.org.apache.bookkeeper.conf.ServerConfiguration;
 import dlshade.org.apache.bookkeeper.meta.AbstractZkLedgerManagerFactory;
@@ -37,7 +38,6 @@ import dlshade.org.apache.bookkeeper.meta.HierarchicalLedgerManagerFactory;
 import dlshade.org.apache.bookkeeper.meta.LayoutManager;
 import dlshade.org.apache.bookkeeper.meta.LedgerLayout;
 import dlshade.org.apache.bookkeeper.meta.LedgerManagerFactory;
-import dlshade.org.apache.bookkeeper.util.ReflectionUtils;
 import java.io.IOException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
